细调怪物文字触发频率和距离，怪物更容易出现黄字，玩家因点击失败被打断闭气添加音效，怪物添加靠近的心跳声音

+++ 2dGame/Assets/GameScripts/UISceneFollower.cs
@@ -7,7 +7,7 @@ using UnityEngine.UI;
 public class UISceneFollower : MonoBehaviour
 {
     public int tag;
-    public int auto_destory_after = 0;
+    public float auto_destory_after = 0;
     public Transform fellowing_obj;
     public Vector3 offset = new Vector3(0, 0, 0);
     private RectTransform rect_transform;
